pyproject.toml
[project]
name = "rbac-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
